+++ Assets/Hotfix/Framework/Manager/FX/Interface/IFXService.cs
@@ -12,11 +12,10 @@ namespace LccHotfix
 
         void ReleaseAll();
 
-        long PlayNormal(string path, Vector3 pos, Quaternion rot, Vector3 scale, float during, bool ignoreTimeScale = false);
+        int PlayNormal(string path, Vector3 pos, Quaternion rot, Vector3 scale, float during, bool ignoreTimeScale = false);
 
-        long PlayNormal(string path, Transform followTrans, float during, bool unit, bool ignoreTimeScale = false);
-
-
-        long PlayLineRender(string path, Transform fromTrans, Transform toTrans, float during, bool follow, bool ignoreTimeScale = false);
+        int PlayNormal(string path, Transform followTrans, float during, bool unit, bool ignoreTimeScale = false);
+        
+        int PlayLineRender(string path, Transform fromTrans, Transform toTrans, float during, bool follow, bool ignoreTimeScale = false);
     }
 }
