+++ Assets/Hotfix/Framework/Utility/Id/IdUtility.cs
@@ -5,7 +5,7 @@ namespace LccHotfix
     public static class IdUtility
     {
         public static int id;
-        public static long GenerateId()
+        public static int GenerateId()
         {
             //https://github.com/leeveel/GeekServer
             return Interlocked.Increment(ref id);
